+++ pages/entry.tsx
@@ -4,8 +4,8 @@ import type { NextPage } from 'next';
 import { EntryTemplate } from '@/components/features/entry/EntryTemplate';
 import { useLoading } from '@/hooks/stateHook/useLoading';
 import { useMeStatus } from '@/hooks/supabaseHook/useMeStatus';
+import { getStripe } from '@/libs/stripe';
 import { createAccount } from '@/models/account/api';
-import { getStripe } from '@/server/stripe/client';
 import { createStripeCheckout } from '@/utils/axios/stripe';
 import { useCounter } from '@mantine/hooks';
 import { useUser } from '@supabase/auth-helpers-react';
